add seatunnel-engine-core module
