+++ generate_client_protocol.sh
@@ -51,8 +51,8 @@ cd $PROTOCOL_DIRECTORY
 
 $PIP3 install -r requirements.txt
 
-$PYTHON generator.py -r $SEATUNNEL_ENGINE_HOME -p $SEATUNNEL_ENGINE_HOME/seatunnel-engine/seatunnel-engine-common/src/main/resources/client-protocol-definition \
--o seatunnel-engine/seatunnel-engine-common/src/main/java/org/apache/seatunnel/engine/protocol/codec \
--n org.apache.seatunnel.engine.protocol.codec --no-binary --no-id-check
+$PYTHON generator.py -r $SEATUNNEL_ENGINE_HOME -p $SEATUNNEL_ENGINE_HOME/seatunnel-engine/seatunnel-engine-core/src/main/resources/client-protocol-definition \
+-o seatunnel-engine/seatunnel-engine-core/src/main/java/org/apache/seatunnel/engine/core/protocol/codec \
+-n org.apache.seatunnel.engine.core.protocol.codec --no-binary --no-id-check
 
 rm -rf $PROTOCOL_DIRECTORY
