+++ seatunnel-engine/pom.xml
@@ -29,5 +29,6 @@
         <module>seatunnel-engine-client</module>
         <module>seatunnel-engine-common</module>
         <module>seatunnel-engine-server</module>
+        <module>seatunnel-engine-core</module>
     </modules>
 </project>
