+++ seatunnel-engine/seatunnel-engine-server/pom.xml
@@ -31,7 +31,7 @@
     <dependencies>
         <dependency>
             <groupId>org.apache.seatunnel</groupId>
-            <artifactId>seatunnel-engine-common</artifactId>
+            <artifactId>seatunnel-engine-core</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
