+++ seatunnel-engine/seatunnel-engine-server/src/main/resources/META-INF/services/com.hazelcast.DataSerializerHook
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.seatunnel.engine.serializable.OperationDataSerializerHook
+org.apache.seatunnel.engine.server.serializable.OperationDataSerializerHook
