+++ seatunnel-engine/seatunnel-engine-server/src/main/resources/META-INF/services/com.hazelcast.client.impl.protocol.MessageTaskFactoryProvider
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.seatunnel.engine.protocol.task.SeaTunnelMessageTaskFactoryProvider
+org.apache.seatunnel.engine.server.protocol.task.SeaTunnelMessageTaskFactoryProvider
